feat: AlarmCard 컴포넌트 구현 및 AlarmList에 적용
