+++ src/components/layout/header/AlarmCard/AlarmCard.module.scss
@@ -1,0 +1,29 @@
+.container {
+  @include column-flexbox($gap: 0.8rem);
+
+  padding: 1.2rem;
+  background: $opacity-white-10;
+  border-radius: 0.8rem;
+
+  &-content {
+    @include flexbox($gap: 0.8rem);
+  }
+}
+
+.content {
+  @include text-style(14);
+}
+
+.delete {
+  @include flexbox;
+
+  flex-shrink: 0;
+  width: 3.2rem;
+  height: 3.2rem;
+}
+
+.created-at {
+  @include text-style(12, $gray10);
+
+  align-self: end;
+}
